+++ cmake/Modules/FindGCSClient.cmake
@@ -104,7 +104,7 @@ elseif(NOT GCSSDK_FOUND)
   ExternalProject_Add(gcssdk-build
     PREFIX ${GCSSDK_PREFIX}
     URL ${GCSSDK_URL}
-    PATCH_COMMAND cp ${CMAKE_CURRENT_SOURCE_DIR}/core/include/misc/tdb_openssl_shim.h 
+    PATCH_COMMAND cp ${CMAKE_CURRENT_SOURCE_DIR}/core/include/misc/tiledb_openssl_shim.h 
                      ${GCSSDK_PREFIX}/src/gcssdk-build/google/cloud/storage &&
                   patch -p1 < ${CMAKE_CURRENT_SOURCE_DIR}/cmake/patches/gcssdk/gcs_ossl.patch
     BUILD_IN_SOURCE 1
