+++ cmake/patches/gcssdk/gcs_ossl.patch
@@ -1,17 +1,17 @@
---- gcssdk-build-orig/google/cloud/storage/hashing_options.cc	2021-02-01 19:02:25.000000000 -0800
-+++ gcssdk-build/google/cloud/storage/hashing_options.cc	2023-03-07 23:16:45.790035203 -0800
+--- gcssdk-build-orig/google/cloud/storage/hashing_options.cc	2021-02-02 08:32:25.000000000 +0530
++++ gcssdk-build/google/cloud/storage/hashing_options.cc	2023-03-14 15:19:40.931760484 +0530
 @@ -14,9 +14,9 @@
  
  #include "google/cloud/storage/hashing_options.h"
  #include "google/cloud/storage/internal/openssl_util.h"
-+#include "google/cloud/storage/tdb_openssl_shim.h"
++#include "google/cloud/storage/tiledb_openssl_shim.h"
  #include "google/cloud/internal/big_endian.h"
  #include <crc32c/crc32c.h>
 -#include <openssl/md5.h>
  #include <cstring>
  
  namespace google {
-@@ -24,13 +24,29 @@
+@@ -24,13 +24,33 @@
  namespace storage {
  inline namespace STORAGE_CLIENT_NS {
  std::string ComputeMD5Hash(std::string const& payload) {
@@ -33,14 +33,18 @@
 +
 +  } else {
 +    EVP_MD_CTX* mdctx;
-+    unsigned char* md5_digest;
-+    unsigned int md5_digest_len = 256; // EVP_MD_get_size(EVP_md5());
++    unsigned int md5_digest_len = 0;
++    if(OpenSSL_version_num() < 0x30000000L ) {
++      md5_digest_len = EVP_MD_size(EVP_sha1());
++    } else {
++      md5_digest_len = EVP_MD_get_size(EVP_sha1());
++    }
++    unsigned char md5_digest[md5_digest_len];
 +
 +    mdctx = EVP_MD_CTX_new();
 +    EVP_DigestInit_ex(mdctx, EVP_md5(), NULL);
 +
 +    EVP_DigestUpdate(mdctx, payload.c_str(), payload.size());
-+    md5_digest = (unsigned char*)OPENSSL_malloc(md5_digest_len);
 +    EVP_DigestFinal_ex(mdctx, md5_digest, &md5_digest_len);
 +    EVP_MD_CTX_free(mdctx);
 +    return internal::Base64Encode(std::string((char*)md5_digest));
@@ -48,13 +52,13 @@
  }
  
  std::string ComputeCrc32cChecksum(std::string const& payload) {
---- gcssdk-build-orig/google/cloud/storage/internal/hash_validator_impl.h	2021-02-01 19:02:25.000000000 -0800
-+++ gcssdk-build/google/cloud/storage/internal/hash_validator_impl.h	2023-03-07 23:16:45.534028956 -0800
+--- gcssdk-build-orig/google/cloud/storage/internal/hash_validator_impl.h	2021-02-02 08:32:25.000000000 +0530
++++ gcssdk-build/google/cloud/storage/internal/hash_validator_impl.h	2023-03-11 21:12:33.842213292 +0530
 @@ -16,8 +16,8 @@
  #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_HASH_VALIDATOR_IMPL_H
  
  #include "google/cloud/storage/internal/hash_validator.h"
-+#include "google/cloud/storage/tdb_openssl_shim.h"
++#include "google/cloud/storage/tiledb_openssl_shim.h"
  #include "google/cloud/storage/version.h"
 -#include <openssl/md5.h>
  
@@ -68,13 +72,13 @@
    std::string received_hash_;
  };
  
---- gcssdk-build-orig/google/cloud/storage/internal/hash_validator_impl.cc	2021-02-01 19:02:25.000000000 -0800
-+++ gcssdk-build/google/cloud/storage/internal/hash_validator_impl.cc	2023-03-07 23:16:45.532028907 -0800
+--- gcssdk-build-orig/google/cloud/storage/internal/hash_validator_impl.cc	2021-02-02 08:32:25.000000000 +0530
++++ gcssdk-build/google/cloud/storage/internal/hash_validator_impl.cc	2023-03-14 15:30:57.680662240 +0530
 @@ -15,6 +15,7 @@
  #include "google/cloud/storage/internal/hash_validator_impl.h"
  #include "google/cloud/storage/internal/openssl_util.h"
  #include "google/cloud/storage/object_metadata.h"
-+#include "google/cloud/storage/tdb_openssl_shim.h"
++#include "google/cloud/storage/tiledb_openssl_shim.h"
  #include "google/cloud/internal/big_endian.h"
  #include <crc32c/crc32c.h>
  
@@ -102,7 +106,7 @@
  }
  
  void MD5HashValidator::ProcessMetadata(ObjectMetadata const& meta) {
-@@ -59,11 +71,22 @@
+@@ -59,11 +71,27 @@
  }
  
  HashValidator::Result MD5HashValidator::Finish() && {
@@ -118,9 +122,14 @@
 +    bool is_mismatch = !received_hash_.empty() && (received_hash_ != computed);
 +    return Result{std::move(received_hash_), std::move(computed), is_mismatch};
 +  } else {
-+    unsigned char* md5_digest;
-+    unsigned int md5_digest_len = 256; // EVP_MD_get_size(EVP_md5());
-+    md5_digest = (unsigned char*)OPENSSL_malloc(md5_digest_len);
++    unsigned int md5_digest_len = 0;
++    if(OpenSSL_version_num() < 0x30000000L ) { 
++      md5_digest_len = EVP_MD_size(EVP_sha1());
++    } else {
++      md5_digest_len = EVP_MD_get_size(EVP_sha1());
++    }   
++    unsigned char md5_digest[md5_digest_len];
++
 +    EVP_DigestFinal_ex(evp_context_, md5_digest, &md5_digest_len);
 +    auto computed = Base64Encode(std::string((char*)md5_digest));
 +    bool is_mismatch = !received_hash_.empty() && (received_hash_ != computed);
@@ -130,22 +139,22 @@
  }
  
  void Crc32cHashValidator::Update(char const* buf, std::size_t n) {
---- gcssdk-build-orig/google/cloud/storage/internal/sha256_hash.cc	2021-02-01 19:02:25.000000000 -0800
-+++ gcssdk-build/google/cloud/storage/internal/sha256_hash.cc	2023-03-07 23:16:45.533028931 -0800
+--- gcssdk-build-orig/google/cloud/storage/internal/sha256_hash.cc	2021-02-02 08:32:25.000000000 +0530
++++ gcssdk-build/google/cloud/storage/internal/sha256_hash.cc	2023-03-14 15:32:06.421885675 +0530
 @@ -13,10 +13,10 @@
  // limitations under the License.
  
  #include "google/cloud/storage/internal/sha256_hash.h"
 -#include <openssl/evp.h>
 -#include <openssl/sha.h>
-+#include "google/cloud/storage/tdb_openssl_shim.h"
++#include "google/cloud/storage/tiledb_openssl_shim.h"
  #include <array>
  
 +#define SHA256_DIGEST_LENGTH    32
  namespace google {
  namespace cloud {
  namespace storage {
-@@ -27,17 +27,31 @@
+@@ -27,17 +27,36 @@
  template <typename Byte,
            typename std::enable_if<sizeof(Byte) == 1, int>::type = 0>
  std::vector<std::uint8_t> Sha256Hash(Byte const* data, std::size_t count) {
@@ -174,13 +183,18 @@
 +    return {hash.begin(), hash.end()};
 +  } else {
 +    EVP_MD_CTX* mdctx = nullptr;
-+    unsigned char* md5_digest;
-+    unsigned int md5_digest_len = 256; //EVP_MD_get_size(EVP_md5());
 +    mdctx = EVP_MD_CTX_new();
 +    EVP_DigestInit_ex(mdctx, EVP_sha256(), NULL);
 +    EVP_DigestUpdate(mdctx, data, count);
 +
-+    md5_digest = (unsigned char*)OPENSSL_malloc(md5_digest_len);
++    unsigned int md5_digest_len = 0;
++    if(OpenSSL_version_num() < 0x30000000L ) { 
++      md5_digest_len = EVP_MD_size(EVP_sha1());
++    } else {
++      md5_digest_len = EVP_MD_get_size(EVP_sha1());
++    }   
++    unsigned char md5_digest[md5_digest_len];
++
 +    EVP_DigestFinal_ex(mdctx, md5_digest, &md5_digest_len);
 +    EVP_MD_CTX_free(mdctx);
 +    return {md5_digest, md5_digest + md5_digest_len};
