+++ core/include/codec/codec_zstd.h
@@ -35,7 +35,6 @@
 #define  __CODEC_ZSTD_H__
 
 #include "codec.h"
-#include <dl_utils.h>
 
 // Function Pointers for ZStd
 #if !defined(ZSTD_EXTERN_DECL)
