+++ core/src/c_api/tiledb_utils.cc
@@ -324,7 +324,7 @@ static int check_file_for_read(TileDB_CTX *tiledb_ctx, std::string filename) {
 }
 
 
-#include "tdb_openssl_shim.h"
+#include "tiledb_openssl_shim.h"
 void print_md5_hash(unsigned char* buffer, size_t length) {
   unsigned char md[MD5_DIGEST_LENGTH];
   MD5(buffer, length, md);
